*: move backfillWorker code to backfilling.go, update logs and comments (#19503)
