+++ ddl/index.go
@@ -16,9 +16,7 @@ package ddl
 import (
 	"context"
 	"math"
-	"strconv"
 	"strings"
-	"sync/atomic"
 	"time"
 
 	"github.com/pingcap/errors"
@@ -28,15 +26,12 @@ import (
 	"github.com/pingcap/parser/model"
 	"github.com/pingcap/parser/mysql"
 	"github.com/pingcap/tidb/config"
-	ddlutil "github.com/pingcap/tidb/ddl/util"
-	"github.com/pingcap/tidb/expression"
 	"github.com/pingcap/tidb/infoschema"
 	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/meta"
 	"github.com/pingcap/tidb/meta/autoid"
 	"github.com/pingcap/tidb/metrics"
 	"github.com/pingcap/tidb/sessionctx"
-	"github.com/pingcap/tidb/sessionctx/variable"
 	"github.com/pingcap/tidb/store/tikv"
 	"github.com/pingcap/tidb/store/tikv/oracle"
 	"github.com/pingcap/tidb/table"
@@ -47,6 +42,7 @@ import (
 	"github.com/pingcap/tidb/util/logutil"
 	decoder "github.com/pingcap/tidb/util/rowDecoder"
 	"github.com/pingcap/tidb/util/timeutil"
+	"github.com/prometheus/client_golang/prometheus"
 	"go.uber.org/zap"
 )
 
@@ -554,7 +550,7 @@ func (w *worker) onCreateIndex(d *ddlCtx, t *meta.Meta, job *model.Job, isPK boo
 				// if timeout, we should return, check for the owner and re-wait job done.
 				return ver, nil
 			}
-			if kv.ErrKeyExists.Equal(err) || errCancelledDDLJob.Equal(err) || errCantDecodeIndex.Equal(err) {
+			if kv.ErrKeyExists.Equal(err) || errCancelledDDLJob.Equal(err) || errCantDecodeRecord.Equal(err) {
 				logutil.BgLogger().Warn("[ddl] run add index job failed, convert job to rollback", zap.String("job", job.String()), zap.Error(err))
 				ver, err = convertAddIdxJob2RollbackJob(t, job, tblInfo, indexInfo, err)
 			}
@@ -795,7 +791,8 @@ type indexRecord struct {
 
 type addIndexWorker struct {
 	*backfillWorker
-	index table.Index
+	index         table.Index
+	metricCounter prometheus.Counter
 
 	// The following attributes are used to reduce memory allocation.
 	defaultVals        []types.Datum
@@ -807,57 +804,21 @@ type addIndexWorker struct {
 	distinctCheckFlags []bool
 }
 
-type reorgIndexTask struct {
-	physicalTableID int64
-	startHandle     kv.Handle
-	endHandle       kv.Handle
-	// endIncluded indicates whether the range include the endHandle.
-	// When the last handle is math.MaxInt64, set endIncluded to true to
-	// tell worker backfilling index of endHandle.
-	endIncluded bool
-}
-
-func (r *reorgIndexTask) String() string {
-	rightParenthesis := ")"
-	if r.endIncluded {
-		rightParenthesis = "]"
-	}
-	return "physicalTableID" + strconv.FormatInt(r.physicalTableID, 10) + "_" + "[" + r.startHandle.String() + "," + r.endHandle.String() + rightParenthesis
-}
-
-type backfillResult struct {
-	addedCount int
-	scanCount  int
-	nextHandle kv.Handle
-	err        error
-}
-
-// backfillTaskContext is the context of the batch adding indices.
-// After finishing the batch adding indices, result in backfillTaskContext will be merged into backfillResult.
-type backfillTaskContext struct {
-	nextHandle kv.Handle
-	done       bool
-	addedCount int
-	scanCount  int
-}
-
-// mergeAddIndexCtxToResult merge partial result in taskCtx into result.
-func mergeAddIndexCtxToResult(taskCtx *backfillTaskContext, result *backfillResult) {
-	result.nextHandle = taskCtx.nextHandle
-	result.addedCount += taskCtx.addedCount
-	result.scanCount += taskCtx.scanCount
-}
-
 func newAddIndexWorker(sessCtx sessionctx.Context, worker *worker, id int, t table.PhysicalTable, indexInfo *model.IndexInfo, decodeColMap map[int64]decoder.Column) *addIndexWorker {
 	index := tables.NewIndex(t.GetPhysicalID(), t.Meta(), indexInfo)
 	rowDecoder := decoder.NewRowDecoder(t, t.WritableCols(), decodeColMap)
 	return &addIndexWorker{
 		backfillWorker: newBackfillWorker(sessCtx, worker, id, t),
 		index:          index,
+		metricCounter:  metrics.BackfillTotalCounter.WithLabelValues("add_idx_speed"),
 		rowDecoder:     rowDecoder,
 		defaultVals:    make([]types.Datum, len(t.WritableCols())),
 		rowMap:         make(map[int64]types.Datum, len(decodeColMap)),
 	}
+}
+
+func (w *addIndexWorker) AddMetricInfo(cnt float64) {
+	w.metricCounter.Add(cnt)
 }
 
 // getIndexRecord gets index columns values from raw binary value row.
@@ -868,7 +829,7 @@ func (w *addIndexWorker) getIndexRecord(handle kv.Handle, recordKey []byte, rawR
 	sysZone := timeutil.SystemLocation()
 	_, err := w.rowDecoder.DecodeAndEvalRowWithMap(w.sessCtx, handle, rawRecord, time.UTC, sysZone, w.rowMap)
 	if err != nil {
-		return nil, errors.Trace(errCantDecodeIndex.GenWithStackByArgs(err))
+		return nil, errors.Trace(errCantDecodeRecord.GenWithStackByArgs("index", err))
 	}
 	idxVal := make([]types.Datum, len(idxInfo.Columns))
 	for j, v := range idxInfo.Columns {
@@ -909,7 +870,7 @@ func (w *addIndexWorker) cleanRowMap() {
 }
 
 // getNextHandle gets next handle of entry that we are going to process.
-func (w *addIndexWorker) getNextHandle(taskRange reorgIndexTask, taskDone bool) (nextHandle kv.Handle) {
+func (w *addIndexWorker) getNextHandle(taskRange reorgBackfillTask, taskDone bool) (nextHandle kv.Handle) {
 	if !taskDone {
 		// The task is not done. So we need to pick the last processed entry's handle and add one.
 		return w.idxRecords[len(w.idxRecords)-1].handle.Next()
@@ -932,7 +893,7 @@ func (w *addIndexWorker) getNextHandle(taskRange reorgIndexTask, taskDone bool) 
 // 2. Next handle of entry that we need to process.
 // 3. Boolean indicates whether the task is done.
 // 4. error occurs in fetchRowColVals. nil if no error occurs.
-func (w *addIndexWorker) fetchRowColVals(txn kv.Transaction, taskRange reorgIndexTask) ([]*indexRecord, kv.Handle, bool, error) {
+func (w *addIndexWorker) fetchRowColVals(txn kv.Transaction, taskRange reorgBackfillTask) ([]*indexRecord, kv.Handle, bool, error) {
 	// TODO: use tableScan to prune columns.
 	w.idxRecords = w.idxRecords[:0]
 	startTime := time.Now()
@@ -976,16 +937,6 @@ func (w *addIndexWorker) fetchRowColVals(txn kv.Transaction, taskRange reorgInde
 
 	logutil.BgLogger().Debug("[ddl] txn fetches handle info", zap.Uint64("txnStartTS", txn.StartTS()), zap.String("taskRange", taskRange.String()), zap.Duration("takeTime", time.Since(startTime)))
 	return w.idxRecords, w.getNextHandle(taskRange, taskDone), taskDone, errors.Trace(err)
-}
-
-func logSlowOperations(elapsed time.Duration, slowMsg string, threshold uint32) {
-	if threshold == 0 {
-		threshold = atomic.LoadUint32(&variable.DDLSlowOprThreshold)
-	}
-
-	if elapsed >= time.Duration(threshold)*time.Millisecond {
-		logutil.BgLogger().Info("[ddl] slow operations", zap.Duration("takeTimes", elapsed), zap.String("msg", slowMsg))
-	}
 }
 
 func (w *addIndexWorker) initBatchCheckBufs(batchCount int) {
@@ -1057,7 +1008,7 @@ func (w *addIndexWorker) batchCheckUniqueKey(txn kv.Transaction, idxRecords []*i
 // indicate that index columns values may changed, index is not allowed to be added, so the txn will rollback and retry.
 // BackfillDataInTxn will add w.batchCnt indices once, default value of w.batchCnt is 128.
 // TODO: make w.batchCnt can be modified by system variable.
-func (w *addIndexWorker) BackfillDataInTxn(handleRange reorgIndexTask) (taskCtx backfillTaskContext, errInTxn error) {
+func (w *addIndexWorker) BackfillDataInTxn(handleRange reorgBackfillTask) (taskCtx backfillTaskContext, errInTxn error) {
 	failpoint.Inject("errorMockPanic", func(val failpoint.Value) {
 		if val.(bool) {
 			panic("panic test")
@@ -1112,401 +1063,14 @@ func (w *addIndexWorker) BackfillDataInTxn(handleRange reorgIndexTask) (taskCtx 
 
 		return nil
 	})
-	logSlowOperations(time.Since(oprStartTime), "BackfillDataInTxn", 3000)
+	logSlowOperations(time.Since(oprStartTime), "AddIndexBackfillDataInTxn", 3000)
 
 	return
-}
-
-var addIndexSpeedCounter = metrics.AddIndexTotalCounter.WithLabelValues("speed")
-
-// handleBackfillTask backfills range [task.startHandle, task.endHandle) handle's index to table.
-func (w *backfillWorker) handleBackfillTask(d *ddlCtx, task *reorgIndexTask, bf backfiller) *backfillResult {
-	handleRange := *task
-	result := &backfillResult{addedCount: 0, nextHandle: handleRange.startHandle, err: nil}
-	lastLogCount := 0
-	lastLogTime := time.Now()
-	startTime := lastLogTime
-
-	for {
-		// Give job chance to be canceled, if we not check it here,
-		// if there is panic in w.BackfillDataInTxn we will never cancel the job.
-		// Because reorgIndexTask may run a long time,
-		// we should check whether this ddl job is still runnable.
-		err := w.ddlWorker.isReorgRunnable(d)
-		if err != nil {
-			result.err = err
-			return result
-		}
-
-		taskCtx, err := bf.BackfillDataInTxn(handleRange)
-		if err != nil {
-			result.err = err
-			return result
-		}
-
-		addIndexSpeedCounter.Add(float64(taskCtx.addedCount))
-		mergeAddIndexCtxToResult(&taskCtx, result)
-		w.ddlWorker.reorgCtx.increaseRowCount(int64(taskCtx.addedCount))
-
-		if num := result.scanCount - lastLogCount; num >= 30000 {
-			lastLogCount = result.scanCount
-			logutil.BgLogger().Info("[ddl] add index worker back fill index", zap.Int("workerID", w.id), zap.Int("addedCount", result.addedCount),
-				zap.Int("scanCount", result.scanCount), zap.String("nextHandle", toString(taskCtx.nextHandle)), zap.Float64("speed(rows/s)", float64(num)/time.Since(lastLogTime).Seconds()))
-			lastLogTime = time.Now()
-		}
-
-		handleRange.startHandle = taskCtx.nextHandle
-		if taskCtx.done {
-			break
-		}
-	}
-	logutil.BgLogger().Info("[ddl] add index worker finish task", zap.Int("workerID", w.id),
-		zap.String("task", task.String()), zap.Int("addedCount", result.addedCount),
-		zap.Int("scanCount", result.scanCount), zap.String("nextHandle", toString(result.nextHandle)),
-		zap.String("takeTime", time.Since(startTime).String()))
-	return result
-}
-
-func (w *backfillWorker) run(d *ddlCtx, bf backfiller) {
-	logutil.BgLogger().Info("[ddl] add index worker start", zap.Int("workerID", w.id))
-	defer func() {
-		w.resultCh <- &backfillResult{err: errReorgPanic}
-	}()
-	defer util.Recover(metrics.LabelDDL, "addIndexWorker.run", nil, false)
-	for {
-		task, more := <-w.taskCh
-		if !more {
-			break
-		}
-
-		logutil.BgLogger().Debug("[ddl] add index worker got task", zap.Int("workerID", w.id), zap.String("task", task.String()))
-		failpoint.Inject("mockAddIndexErr", func() {
-			if w.id == 0 {
-				result := &backfillResult{addedCount: 0, nextHandle: nil, err: errors.Errorf("mock add index error")}
-				w.resultCh <- result
-				failpoint.Continue()
-			}
-		})
-
-		// Dynamic change batch size.
-		w.batchCnt = int(variable.GetDDLReorgBatchSize())
-		result := w.handleBackfillTask(d, task, bf)
-		w.resultCh <- result
-	}
-	logutil.BgLogger().Info("[ddl] add index worker exit", zap.Int("workerID", w.id))
-}
-
-func makeupDecodeColMap(sessCtx sessionctx.Context, t table.Table) (map[int64]decoder.Column, error) {
-	dbName := model.NewCIStr(sessCtx.GetSessionVars().CurrentDB)
-	writableColInfos := make([]*model.ColumnInfo, 0, len(t.WritableCols()))
-	for _, col := range t.WritableCols() {
-		writableColInfos = append(writableColInfos, col.ColumnInfo)
-	}
-	exprCols, _, err := expression.ColumnInfos2ColumnsAndNames(sessCtx, dbName, t.Meta().Name, writableColInfos, t.Meta())
-	if err != nil {
-		return nil, err
-	}
-	mockSchema := expression.NewSchema(exprCols...)
-
-	decodeColMap := decoder.BuildFullDecodeColMap(t.WritableCols(), mockSchema)
-
-	return decodeColMap, nil
-}
-
-// splitTableRanges uses PD region's key ranges to split the backfilling table key range space,
-// to speed up adding index in table with disperse handle.
-// The `t` should be a non-partitioned table or a partition.
-func splitTableRanges(t table.PhysicalTable, store kv.Storage, startHandle, endHandle kv.Handle) ([]kv.KeyRange, error) {
-	startRecordKey := t.RecordKey(startHandle)
-	endRecordKey := t.RecordKey(endHandle)
-
-	logutil.BgLogger().Info("[ddl] split table range from PD", zap.Int64("physicalTableID", t.GetPhysicalID()),
-		zap.String("startHandle", toString(startHandle)), zap.String("endHandle", toString(endHandle)))
-	kvRange := kv.KeyRange{StartKey: startRecordKey, EndKey: endRecordKey}
-	s, ok := store.(tikv.Storage)
-	if !ok {
-		// Only support split ranges in tikv.Storage now.
-		return []kv.KeyRange{kvRange}, nil
-	}
-
-	maxSleep := 10000 // ms
-	bo := tikv.NewBackofferWithVars(context.Background(), maxSleep, nil)
-	ranges, err := tikv.SplitRegionRanges(bo, s.GetRegionCache(), []kv.KeyRange{kvRange})
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	if len(ranges) == 0 {
-		return nil, errors.Trace(errInvalidSplitRegionRanges)
-	}
-	return ranges, nil
-}
-
-func decodeHandleRange(keyRange kv.KeyRange) (kv.Handle, kv.Handle, error) {
-	startHandle, err := tablecodec.DecodeRowKey(keyRange.StartKey)
-	if err != nil {
-		return nil, nil, errors.Trace(err)
-	}
-	endHandle, err := tablecodec.DecodeRowKey(keyRange.EndKey)
-	if err != nil {
-		return nil, nil, errors.Trace(err)
-	}
-
-	return startHandle, endHandle, nil
-}
-
-func (w *worker) waitTaskResults(workers []*backfillWorker, taskCnt int, totalAddedCount *int64, startHandle kv.Handle) (kv.Handle, int64, error) {
-	var (
-		addedCount int64
-		nextHandle = startHandle
-		firstErr   error
-	)
-	for i := 0; i < taskCnt; i++ {
-		worker := workers[i]
-		result := <-worker.resultCh
-		if firstErr == nil && result.err != nil {
-			firstErr = result.err
-			// We should wait all working workers exits, any way.
-			continue
-		}
-
-		if result.err != nil {
-			logutil.BgLogger().Warn("[ddl] add index worker failed", zap.Int("workerID", worker.id),
-				zap.Error(result.err))
-		}
-
-		if firstErr == nil {
-			*totalAddedCount += int64(result.addedCount)
-			addedCount += int64(result.addedCount)
-			nextHandle = result.nextHandle
-		}
-	}
-
-	return nextHandle, addedCount, errors.Trace(firstErr)
-}
-
-// handleReorgTasks sends tasks to workers, and waits for all the running workers to return results,
-// there are taskCnt running workers.
-func (w *worker) handleReorgTasks(reorgInfo *reorgInfo, totalAddedCount *int64, workers []*backfillWorker, batchTasks []*reorgIndexTask) error {
-	for i, task := range batchTasks {
-		workers[i].taskCh <- task
-	}
-
-	startHandle := batchTasks[0].startHandle
-	taskCnt := len(batchTasks)
-	startTime := time.Now()
-	nextHandle, taskAddedCount, err := w.waitTaskResults(workers, taskCnt, totalAddedCount, startHandle)
-	elapsedTime := time.Since(startTime)
-	if err == nil {
-		err = w.isReorgRunnable(reorgInfo.d)
-	}
-
-	if err != nil {
-		// update the reorg handle that has been processed.
-		err1 := kv.RunInNewTxn(reorgInfo.d.store, true, func(txn kv.Transaction) error {
-			return errors.Trace(reorgInfo.UpdateReorgMeta(txn, nextHandle, reorgInfo.EndHandle, reorgInfo.PhysicalTableID))
-		})
-		metrics.BatchAddIdxHistogram.WithLabelValues(metrics.LblError).Observe(elapsedTime.Seconds())
-		logutil.BgLogger().Warn("[ddl] add index worker handle batch tasks failed",
-			zap.Int64("totalAddedCount", *totalAddedCount), zap.String("startHandle", toString(startHandle)),
-			zap.String("nextHandle", toString(nextHandle)), zap.Int64("batchAddedCount", taskAddedCount),
-			zap.String("taskFailedError", err.Error()), zap.String("takeTime", elapsedTime.String()),
-			zap.NamedError("updateHandleError", err1))
-		return errors.Trace(err)
-	}
-
-	// nextHandle will be updated periodically in runReorgJob, so no need to update it here.
-	w.reorgCtx.setNextHandle(nextHandle)
-	metrics.BatchAddIdxHistogram.WithLabelValues(metrics.LblOK).Observe(elapsedTime.Seconds())
-	logutil.BgLogger().Info("[ddl] add index worker handle batch tasks successful", zap.Int64("totalAddedCount", *totalAddedCount), zap.String("startHandle", toString(startHandle)),
-		zap.String("nextHandle", toString(nextHandle)), zap.Int64("batchAddedCount", taskAddedCount), zap.String("takeTime", elapsedTime.String()))
-	return nil
-}
-
-// sendRangeTaskToWorkers sends tasks to workers, and returns remaining kvRanges that is not handled.
-func (w *worker) sendRangeTaskToWorkers(workers []*backfillWorker, reorgInfo *reorgInfo,
-	totalAddedCount *int64, kvRanges []kv.KeyRange, globalEndHandle kv.Handle) ([]kv.KeyRange, error) {
-	batchTasks := make([]*reorgIndexTask, 0, len(workers))
-	physicalTableID := reorgInfo.PhysicalTableID
-
-	// Build reorg indices tasks.
-	for _, keyRange := range kvRanges {
-		startHandle, endHandle, err := decodeHandleRange(keyRange)
-		if err != nil {
-			return nil, errors.Trace(err)
-		}
-
-		endIncluded := false
-		if endHandle.Equal(globalEndHandle) {
-			endIncluded = true
-		}
-		task := &reorgIndexTask{physicalTableID, startHandle, endHandle, endIncluded}
-		batchTasks = append(batchTasks, task)
-
-		if len(batchTasks) >= len(workers) {
-			break
-		}
-	}
-
-	if len(batchTasks) == 0 {
-		return nil, nil
-	}
-
-	// Wait tasks finish.
-	err := w.handleReorgTasks(reorgInfo, totalAddedCount, workers, batchTasks)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-
-	if len(batchTasks) < len(kvRanges) {
-		// there are kvRanges not handled.
-		remains := kvRanges[len(batchTasks):]
-		return remains, nil
-	}
-
-	return nil, nil
-}
-
-var (
-	// TestCheckWorkerNumCh use for test adjust add index worker.
-	TestCheckWorkerNumCh = make(chan struct{})
-	// TestCheckWorkerNumber use for test adjust add index worker.
-	TestCheckWorkerNumber = int32(16)
-)
-
-func loadDDLReorgVars(w *worker) error {
-	// Get sessionctx from context resource pool.
-	var ctx sessionctx.Context
-	ctx, err := w.sessPool.get()
-	if err != nil {
-		return errors.Trace(err)
-	}
-	defer w.sessPool.put(ctx)
-	return ddlutil.LoadDDLReorgVars(ctx)
 }
 
 func (w *worker) addPhysicalTableIndex(t table.PhysicalTable, indexInfo *model.IndexInfo, reorgInfo *reorgInfo) error {
 	logutil.BgLogger().Info("[ddl] start to add table index", zap.String("job", reorgInfo.Job.String()), zap.String("reorgInfo", reorgInfo.String()))
 	return w.writePhysicalTableRecord(t.(table.PhysicalTable), typeAddIndexWorker, indexInfo, nil, nil, reorgInfo)
-}
-
-// writePhysicalTableRecord handles the "add index" or "modify/change column" reorganization state for a non-partitioned table or a partition.
-// For a partitioned table, it should be handled partition by partition.
-//
-// How to "add index" or "update column value" in reorganization state?
-// Concurrently process the @@tidb_ddl_reorg_worker_cnt tasks. Each task deals with a handle range of the index/row record.
-// The handle range is split from PD regions now. Each worker deal with a region table key range one time.
-// Each handle range by estimation, concurrent processing needs to perform after the handle range has been acquired.
-// The operation flow is as follows:
-//	1. Open numbers of defaultWorkers goroutines.
-//	2. Split table key range from PD regions.
-//	3. Send tasks to running workers by workers's task channel. Each task deals with a region key ranges.
-//	4. Wait all these running tasks finished, then continue to step 3, until all tasks is done.
-// The above operations are completed in a transaction.
-// Finally, update the concurrent processing of the total number of rows, and store the completed handle value.
-func (w *worker) writePhysicalTableRecord(t table.PhysicalTable, bfWorkerType backfillWorkerType,
-	indexInfo *model.IndexInfo, oldColInfo, colInfo *model.ColumnInfo, reorgInfo *reorgInfo) error {
-	job := reorgInfo.Job
-	totalAddedCount := job.GetRowCount()
-
-	if err := w.isReorgRunnable(reorgInfo.d); err != nil {
-		return errors.Trace(err)
-	}
-
-	failpoint.Inject("MockCaseWhenParseFailure", func(val failpoint.Value) {
-		if val.(bool) {
-			failpoint.Return(errors.New("job.ErrCount:" + strconv.Itoa(int(job.ErrorCount)) + ", mock unknown type: ast.whenClause."))
-		}
-	})
-
-	startHandle, endHandle := reorgInfo.StartHandle, reorgInfo.EndHandle
-	sessCtx := newContext(reorgInfo.d.store)
-	decodeColMap, err := makeupDecodeColMap(sessCtx, t)
-	if err != nil {
-		return errors.Trace(err)
-	}
-
-	if startHandle == nil && endHandle == nil {
-		return nil
-	}
-
-	// variable.ddlReorgWorkerCounter can be modified by system variable "tidb_ddl_reorg_worker_cnt".
-	workerCnt := variable.GetDDLReorgWorkerCounter()
-	backfillWorkers := make([]*backfillWorker, 0, workerCnt)
-	defer func() {
-		closeBackfillWorkers(backfillWorkers)
-	}()
-
-	for {
-		kvRanges, err := splitTableRanges(t, reorgInfo.d.store, startHandle, endHandle)
-		if err != nil {
-			return errors.Trace(err)
-		}
-
-		// For dynamic adjust add index worker number.
-		if err := loadDDLReorgVars(w); err != nil {
-			logutil.BgLogger().Error("[ddl] load DDL reorganization variable failed", zap.Error(err))
-		}
-		workerCnt = variable.GetDDLReorgWorkerCounter()
-		// If only have 1 range, we can only start 1 worker.
-		if len(kvRanges) < int(workerCnt) {
-			workerCnt = int32(len(kvRanges))
-		}
-		// Enlarge the worker size.
-		for i := len(backfillWorkers); i < int(workerCnt); i++ {
-			sessCtx := newContext(reorgInfo.d.store)
-			sessCtx.GetSessionVars().StmtCtx.IsDDLJobInQueue = true
-
-			if bfWorkerType == typeAddIndexWorker {
-				idxWorker := newAddIndexWorker(sessCtx, w, i, t, indexInfo, decodeColMap)
-				idxWorker.priority = job.Priority
-				backfillWorkers = append(backfillWorkers, idxWorker.backfillWorker)
-				go idxWorker.backfillWorker.run(reorgInfo.d, idxWorker)
-			} else {
-				updateWorker := newUpdateColumnWorker(sessCtx, w, i, t, oldColInfo, colInfo, decodeColMap)
-				updateWorker.priority = job.Priority
-				backfillWorkers = append(backfillWorkers, updateWorker.backfillWorker)
-				go updateWorker.backfillWorker.run(reorgInfo.d, updateWorker)
-			}
-		}
-		// Shrink the worker size.
-		if len(backfillWorkers) > int(workerCnt) {
-			workers := backfillWorkers[workerCnt:]
-			backfillWorkers = backfillWorkers[:workerCnt]
-			closeBackfillWorkers(workers)
-		}
-
-		failpoint.Inject("checkIndexWorkerNum", func(val failpoint.Value) {
-			if val.(bool) {
-				num := int(atomic.LoadInt32(&TestCheckWorkerNumber))
-				if num != 0 {
-					if num > len(kvRanges) {
-						if len(backfillWorkers) != len(kvRanges) {
-							failpoint.Return(errors.Errorf("check index worker num error, len kv ranges is: %v, check index worker num is: %v, actual index num is: %v", len(kvRanges), num, len(backfillWorkers)))
-						}
-					} else if num != len(backfillWorkers) {
-						failpoint.Return(errors.Errorf("check index worker num error, len kv ranges is: %v, check index worker num is: %v, actual index num is: %v", len(kvRanges), num, len(backfillWorkers)))
-					}
-					TestCheckWorkerNumCh <- struct{}{}
-				}
-			}
-		})
-
-		logutil.BgLogger().Info("[ddl] start add index workers to reorg index", zap.Int("workerCnt", len(backfillWorkers)),
-			zap.Int("regionCnt", len(kvRanges)), zap.String("startHandle", toString(startHandle)), zap.String("endHandle", toString(endHandle)))
-		remains, err := w.sendRangeTaskToWorkers(backfillWorkers, reorgInfo, &totalAddedCount, kvRanges, endHandle)
-		if err != nil {
-			return errors.Trace(err)
-		}
-
-		if len(remains) == 0 {
-			break
-		}
-		startHandle, _, err = decodeHandleRange(remains[0])
-		if err != nil {
-			return errors.Trace(err)
-		}
-	}
-	return nil
 }
 
 // addTableIndex handles the add index reorganization state for a table.
@@ -1600,74 +1164,6 @@ func findNextPartitionID(currentPartition int64, defs []model.PartitionDefinitio
 func allocateIndexID(tblInfo *model.TableInfo) int64 {
 	tblInfo.MaxIndexID++
 	return tblInfo.MaxIndexID
-}
-
-// recordIterFunc is used for low-level record iteration.
-type recordIterFunc func(h kv.Handle, rowKey kv.Key, rawRecord []byte) (more bool, err error)
-
-func iterateSnapshotRows(store kv.Storage, priority int, t table.Table, version uint64, startHandle kv.Handle, endHandle kv.Handle, endIncluded bool, fn recordIterFunc) error {
-	var firstKey kv.Key
-	if startHandle == nil {
-		firstKey = t.RecordPrefix()
-	} else {
-		firstKey = t.RecordKey(startHandle)
-	}
-
-	var upperBound kv.Key
-	if endHandle == nil {
-		upperBound = t.RecordPrefix().PrefixNext()
-	} else {
-		if endIncluded {
-			if endHandle.IsInt() && endHandle.IntValue() == math.MaxInt64 {
-				upperBound = t.RecordKey(endHandle).PrefixNext()
-			} else {
-				upperBound = t.RecordKey(endHandle.Next())
-			}
-		} else {
-			upperBound = t.RecordKey(endHandle)
-		}
-	}
-
-	ver := kv.Version{Ver: version}
-	snap, err := store.GetSnapshot(ver)
-	snap.SetOption(kv.Priority, priority)
-	if err != nil {
-		return errors.Trace(err)
-	}
-
-	it, err := snap.Iter(firstKey, upperBound)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	defer it.Close()
-
-	for it.Valid() {
-		if !it.Key().HasPrefix(t.RecordPrefix()) {
-			break
-		}
-
-		var handle kv.Handle
-		handle, err = tablecodec.DecodeRowKey(it.Key())
-		if err != nil {
-			return errors.Trace(err)
-		}
-		rk := t.RecordKey(handle)
-
-		more, err := fn(handle, rk, it.Value())
-		if !more || err != nil {
-			return errors.Trace(err)
-		}
-
-		err = kv.NextUntil(it, util.RowKeyPrefixFilter(rk))
-		if err != nil {
-			if kv.ErrNotExist.Equal(err) {
-				break
-			}
-			return errors.Trace(err)
-		}
-	}
-
-	return nil
 }
 
 func getIndexInfoByNameAndColumn(oldTableInfo *model.TableInfo, newOne *model.IndexInfo) *model.IndexInfo {
