+++ errno/errcode.go
@@ -1049,7 +1049,7 @@ const (
 	// Error codes used by TiDB ddl package
 	ErrUnsupportedDDLOperation            = 8200
 	ErrNotOwner                           = 8201
-	ErrCantDecodeIndex                    = 8202
+	ErrCantDecodeRecord                   = 8202
 	ErrInvalidDDLWorker                   = 8203
 	ErrInvalidDDLJob                      = 8204
 	ErrInvalidDDLJobFlag                  = 8205
